+++ lib/flutter_async.dart
@@ -1,7 +1,9 @@
 library flutter_async;
 
-export 'package:flutter_async/src/widgets/async_button/async_button_base.dart';
-export 'package:flutter_async/src/async_extension.dart';
+export 'src/widgets/async_button/async_button_base.dart';
+export 'src/async_extension.dart';
 export 'src/widgets/async_button/async_button.dart';
 export 'src/widgets/async_builder/async_builder.dart';
 export 'src/async_controller.dart';
+export 'src/widgets/async/inherited_async.dart';
+export 'src/widgets/async/async_config.dart';